+++ HISTORY.rst
@@ -92,6 +92,8 @@
 * Expose Django's LOGOUT_REDIRECT_URL setting.
 * Move current user views from the common app to the user
   management app.
+* Move the purge permission logic to the StorePermission
+  manager.
 
 3.1.11 (2019-04-XX)
 ===================